server/app.py
#server/app.py
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

"""
FastAPI application for the Veritasops Environment.

This module creates an HTTP server that exposes the VeritasopsEnvironment
over HTTP and WebSocket endpoints, compatible with EnvClient.

Endpoints:
    - POST /reset: Reset the environment
    - POST /step: Execute an action
    - GET /state: Get current environment state
    - GET /schema: Get action/observation schemas
    - WS /ws: WebSocket endpoint for persistent sessions

Usage:
    # Development (with auto-reload):
    uvicorn server.app:app --reload --host 0.0.0.0 --port 8000

    # Production:
    uvicorn server.app:app --host 0.0.0.0 --port 8000 --workers 4

    # Or run directly:
    python -m server.app
"""

from openenv.core.env_server.http_server import create_app
from models import VeritasopsAction, VeritasopsObservation
from server.veritasops_environment import VeritasopsEnvironment


app = create_app(
    VeritasopsEnvironment,
    VeritasopsAction,
    VeritasopsObservation,
    env_name="veritasops",
    max_concurrent_envs=4,
)


def main():
    """
    OpenEnv-compatible callable entry point.
    """
    import argparse
    import uvicorn

    parser = argparse.ArgumentParser(description="Run VeritasOps server")
    parser.add_argument("--host", type=str, default="0.0.0.0")
    parser.add_argument("--port", type=int, default=8000)
    args = parser.parse_args()

    uvicorn.run(app, host=args.host, port=args.port)


if __name__ == "__main__":
    main()