pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-veritasops"
version = "0.1.0"
description = "VeritasOps environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi",
    "uvicorn",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
py-modules = [
    "client",
    "models",
    "tasks",
    "simulator",
    "rewards",
    "utils",
    "grader",
    "benchmark",
    "inference",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]